+++ compiler-core/src/metadata/tests.rs
@@ -1,3 +1,4 @@
+use hexpm::version::Version;
 use rand::Rng;
 use type_::{AccessorsMap, FieldMap, RecordAccessor};
 
@@ -62,6 +63,7 @@ fn constant_module(constant: TypedConstant) -> ModuleInterface {
         .into(),
         line_numbers: LineNumbers::new(""),
         src_path: "some_path".into(),
+        minimum_required_version: Version::new(1, 0, 0),
     }
 }
 
@@ -94,6 +96,7 @@ fn empty_module() {
         accessors: HashMap::new(),
         line_numbers: LineNumbers::new(""),
         src_path: "some_path".into(),
+        minimum_required_version: Version::new(1, 0, 0),
     };
     assert_eq!(roundtrip(&module), module);
 }
@@ -116,6 +119,7 @@ fn with_line_numbers() {
         const c = 3",
         ),
         src_path: "some_path".into(),
+        minimum_required_version: Version::new(1, 0, 0),
     };
     assert_eq!(roundtrip(&module), module);
 }
@@ -146,6 +150,7 @@ fn module_with_private_type() {
         accessors: HashMap::new(),
         line_numbers: LineNumbers::new(""),
         src_path: "some_path".into(),
+        minimum_required_version: Version::new(1, 0, 0),
     };
     assert_eq!(roundtrip(&module), module);
 }
@@ -176,6 +181,7 @@ fn module_with_app_type() {
         accessors: HashMap::new(),
         line_numbers: LineNumbers::new(""),
         src_path: "some_path".into(),
+        minimum_required_version: Version::new(1, 0, 0),
     };
     assert_eq!(roundtrip(&module), module);
 }
@@ -206,6 +212,7 @@ fn module_with_fn_type() {
         accessors: HashMap::new(),
         line_numbers: LineNumbers::new(""),
         src_path: "some_path".into(),
+        minimum_required_version: Version::new(1, 0, 0),
     };
     assert_eq!(roundtrip(&module), module);
 }
@@ -236,6 +243,7 @@ fn module_with_tuple_type() {
         accessors: HashMap::new(),
         line_numbers: LineNumbers::new(""),
         src_path: "some_path".into(),
+        minimum_required_version: Version::new(1, 0, 0),
     };
     assert_eq!(roundtrip(&module), module);
 }
@@ -272,6 +280,7 @@ fn module_with_generic_type() {
             accessors: HashMap::new(),
             line_numbers: LineNumbers::new(""),
             src_path: "some_path".into(),
+            minimum_required_version: Version::new(1, 0, 0),
         }
     }
 
@@ -293,7 +302,7 @@ fn module_with_type_links() {
             types: [(
                 "SomeType".into(),
                 TypeConstructor {
-                    type_: type_,
+                    type_,
                     publicity: Publicity::Public,
                     origin: Default::default(),
                     module: "a".into(),
@@ -308,6 +317,7 @@ fn module_with_type_links() {
             accessors: HashMap::new(),
             line_numbers: LineNumbers::new(""),
             src_path: "some_path".into(),
+            minimum_required_version: Version::new(1, 0, 0),
         }
     }
 
@@ -329,7 +339,7 @@ fn module_with_type_constructor_documentation() {
             types: [(
                 "SomeType".into(),
                 TypeConstructor {
-                    type_: type_,
+                    type_,
                     publicity: Publicity::Public,
                     origin: Default::default(),
                     module: "a".into(),
@@ -344,6 +354,7 @@ fn module_with_type_constructor_documentation() {
             accessors: HashMap::new(),
             line_numbers: LineNumbers::new(""),
             src_path: "some_path".into(),
+            minimum_required_version: Version::new(1, 0, 0),
         }
     }
 
@@ -365,7 +376,7 @@ fn module_with_type_constructor_origin() {
             types: [(
                 "SomeType".into(),
                 TypeConstructor {
-                    type_: type_,
+                    type_,
                     publicity: Publicity::Public,
                     origin: SrcSpan {
                         start: 535,
@@ -383,6 +394,7 @@ fn module_with_type_constructor_origin() {
             accessors: HashMap::new(),
             line_numbers: LineNumbers::new(""),
             src_path: "some_path".into(),
+            minimum_required_version: Version::new(1, 0, 0),
         }
     }
 
@@ -413,6 +425,7 @@ fn module_type_to_constructors_mapping() {
         values: HashMap::new(),
         line_numbers: LineNumbers::new(""),
         src_path: "some_path".into(),
+        minimum_required_version: Version::new(1, 0, 0),
     };
 
     assert_eq!(roundtrip(&module), module);
@@ -458,6 +471,7 @@ fn module_fn_value() {
         .into(),
         line_numbers: LineNumbers::new(""),
         src_path: "some_path".into(),
+        minimum_required_version: Version::new(1, 0, 0),
     };
     assert_eq!(roundtrip(&module), module);
 }
@@ -504,6 +518,7 @@ fn deprecated_module_fn_value() {
         .into(),
         line_numbers: LineNumbers::new(""),
         src_path: "some_path".into(),
+        minimum_required_version: Version::new(1, 0, 0),
     };
     assert_eq!(roundtrip(&module), module);
 }
@@ -548,6 +563,7 @@ fn private_module_fn_value() {
         .into(),
         line_numbers: LineNumbers::new(""),
         src_path: "some_path".into(),
+        minimum_required_version: Version::new(1, 0, 0),
     };
 
     assert_eq!(roundtrip(&module), module);
@@ -594,6 +610,7 @@ fn module_fn_value_regression() {
         .into(),
         line_numbers: LineNumbers::new(""),
         src_path: "some_path".into(),
+        minimum_required_version: Version::new(1, 0, 0),
     };
 
     assert_eq!(roundtrip(&module), module);
@@ -639,6 +656,7 @@ fn module_fn_value_with_field_map() {
         .into(),
         line_numbers: LineNumbers::new(""),
         src_path: "some_path".into(),
+        minimum_required_version: Version::new(1, 0, 0),
     };
 
     assert_eq!(roundtrip(&module), module);
@@ -681,6 +699,7 @@ fn record_value() {
         .into(),
         line_numbers: LineNumbers::new(""),
         src_path: "some_path".into(),
+        minimum_required_version: Version::new(1, 0, 0),
     };
 
     assert_eq!(roundtrip(&module), module);
@@ -726,6 +745,7 @@ fn record_value_with_field_map() {
         .into(),
         line_numbers: LineNumbers::new(""),
         src_path: "some_path".into(),
+        minimum_required_version: Version::new(1, 0, 0),
     };
 
     assert_eq!(roundtrip(&module), module);
@@ -789,6 +809,7 @@ fn accessors() {
         .into(),
         line_numbers: LineNumbers::new(""),
         src_path: "some_path".into(),
+        minimum_required_version: Version::new(1, 0, 0),
     };
 
     assert_eq!(roundtrip(&module), module);
@@ -1002,6 +1023,7 @@ fn constant_var() {
         .into(),
         line_numbers: LineNumbers::new(""),
         src_path: "some_path".into(),
+        minimum_required_version: Version::new(1, 0, 0),
     };
 
     assert_eq!(roundtrip(&module), module);
@@ -1191,6 +1213,7 @@ fn deprecated_type() {
         accessors: HashMap::new(),
         line_numbers: LineNumbers::new(""),
         src_path: "some_path".into(),
+        minimum_required_version: Version::new(1, 0, 0),
     };
     assert_eq!(roundtrip(&module), module);
 }
@@ -1235,6 +1258,7 @@ fn module_fn_value_with_external_implementations() {
         .into(),
         line_numbers: LineNumbers::new(""),
         src_path: "some_path".into(),
+        minimum_required_version: Version::new(1, 0, 0),
     };
 
     assert_eq!(roundtrip(&module), module);
@@ -1254,7 +1278,9 @@ fn internal_module_fn() {
         values: [(
             "one".into(),
             ValueConstructor {
-                publicity: Publicity::Internal,
+                publicity: Publicity::Internal {
+                    attribute_location: None,
+                },
                 deprecation: Deprecation::NotDeprecated,
                 type_: type_::int(),
                 variant: ValueConstructorVariant::ModuleFn {
@@ -1280,6 +1306,58 @@ fn internal_module_fn() {
         .into(),
         line_numbers: LineNumbers::new(""),
         src_path: "some_path".into(),
+        minimum_required_version: Version::new(1, 0, 0),
+    };
+
+    assert_eq!(roundtrip(&module), module);
+}
+
+#[test]
+fn internal_annotated_module_fn() {
+    let module = ModuleInterface {
+        warnings: vec![],
+        is_internal: false,
+        package: "some_package".into(),
+        origin: Origin::Src,
+        name: "a/b/c".into(),
+        types: HashMap::new(),
+        types_value_constructors: HashMap::new(),
+        accessors: HashMap::new(),
+        values: [(
+            "one".into(),
+            ValueConstructor {
+                publicity: Publicity::Internal {
+                    attribute_location: Some(SrcSpan {
+                        start: 11,
+                        end: 111,
+                    }),
+                },
+                deprecation: Deprecation::NotDeprecated,
+                type_: type_::int(),
+                variant: ValueConstructorVariant::ModuleFn {
+                    documentation: Some("wabble!".into()),
+                    name: "one".into(),
+                    field_map: None,
+                    module: "a".into(),
+                    arity: 5,
+                    location: SrcSpan {
+                        start: 52,
+                        end: 1100,
+                    },
+                    implementations: Implementations {
+                        gleam: false,
+                        uses_erlang_externals: true,
+                        uses_javascript_externals: true,
+                        can_run_on_erlang: true,
+                        can_run_on_javascript: true,
+                    },
+                },
+            },
+        )]
+        .into(),
+        line_numbers: LineNumbers::new(""),
+        src_path: "some_path".into(),
+        minimum_required_version: Version::new(1, 0, 0),
     };
 
     assert_eq!(roundtrip(&module), module);
@@ -1319,6 +1397,7 @@ fn type_variable_ids_in_constructors_are_shared() {
         values: [].into(),
         line_numbers: LineNumbers::new(""),
         src_path: "some_path".into(),
+        minimum_required_version: Version::new(1, 0, 0),
     };
 
     let expected = HashMap::from([(