+++ CHANGELOG.md
@@ -233,8 +233,43 @@
   Whereas before, it would suggest `_` as the only missing pattern.
   ([Surya Rose](https://github.com/GearsDatapacks))
 
-- Improve error message for using @external with unknown target
+- Improve error message for using `@external` with unknown target
   ([Jiangda Wang](https://github.com/frank-iii))
+
+- Improved error title when using an unknown module value.
+  ([Giacomo Cavalieri](https://github.com/giacomocavalieri))
+
+- The compiler now shows an helpful error message if you try writing an `if`
+  expression instead of a case. For example, this code:
+
+  ```gleam
+  pub fn main() {
+    let a = if wibble {
+      1
+    }
+  }
+  ```
+
+  Results in the following error:
+
+  ```txt
+  error: Syntax error
+    ┌─ /src/parse/error.gleam:3:11
+    │
+  3 │   let a = if wibble {
+    │           ^^ Gleam doesn't have if expressions
+
+  If you want to write a conditional expression you can use a `case`:
+
+      case condition {
+        True -> todo
+        False -> todo
+      }
+
+  See: https://tour.gleam.run/flow-control/case-expressions/
+  ```
+
+  ([Giacomo Cavalieri](https://github.com/giacomocavalieri))
 
 ### Formatter
 
@@ -363,7 +398,8 @@
   unused.
   ([Giacomo Cavalieri](https://github.com/giacomocavalieri))
 
-- The Language Server now correctly shows completions for values in the Gleam prelude.
+- The Language Server now correctly shows completions for values in the Gleam
+  prelude.
   ([Surya Rose](https://github.com/GearsDatapacks))
 
 ## v1.4.1 - 2024-08-04